Change: UINT to uint64_t and add a funy logo

+++ engine/skiplist.h
@@ -2,7 +2,6 @@
 #define _SKIPLIST_H
 
 #include "util.h"
-#include "platform.h"
 
 #define MAXLEVEL (15)
 #define SKIP_KSIZE (256)
@@ -11,7 +10,7 @@ typedef enum {ADD,DEL} OPT;
 
 struct skipnode{
     char key[SKIP_KSIZE];
-	UINT val;
+	uint64_t val;
 	OPT opt;                   
     struct skipnode *forward[1]; 
 	struct skipnode *next;
@@ -28,7 +27,7 @@ struct skiplist{
 
 struct skiplist *skiplist_new(size_t size);
 void skiplist_init(struct skiplist *list);
-int skiplist_insert(struct skiplist *list, struct slice *sk, UINT val, OPT opt);
+int skiplist_insert(struct skiplist *list, struct slice *sk, uint64_t val, OPT opt);
 int skiplist_notfull(struct skiplist *list);
 void skiplist_dump(struct skiplist *list);
 void skiplist_free(struct skiplist *list);
